renamed research and inner research to delegate research and research agents
